+++ src/mash/CommandSketch.cpp
@@ -117,7 +117,7 @@ int CommandSketch::run() const
     
 	for ( int i = 0; i < sketch.getReferenceCount(); i++ )
 	{
-		int length = sketch.getReference(i).length;
+		uint64_t length = sketch.getReference(i).length;
 		
 		if ( length > lengthThreshold )
 		{
